TK176193T6--2/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
altair==4.2.2
amqp==5.2.0
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
aspose-words==24.5.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.3.0
autograd==1.7.0
autoray==0.6.12
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.1
blessed==1.20.0
blinker==1.7.0
build==1.2.2.post1
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.4
configobj==5.0.9
configparser==7.1.0
contourpy==1.2.1
crispy-bootstrap5==2024.2
crypto==1.4.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==4.4.2
Deprecated==1.2.15
distlib==0.3.8
Django==5.1.3
django-crispy-forms==2.3
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
docx2txt==0.8
durationpy==0.9
efficientnet_pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.5.0
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
gevent==24.11.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.1
googlemaps==4.10.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jinxed==1.3.0
joblib==1.3.2
json2html==1.3.0
jsonify==0.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.5
kombu==5.4.2
kubernetes==31.0.0
langchain==0.0.284
langchain-experimental==0.0.25
langsmith==0.0.92
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.3.0
looseversion==1.3.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==5.0.1
monai==1.4.0
monotonic==1.6
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
Naked==0.1.32
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.3.2
nipype==1.8.6
nltk==3.9.1
ntru==0.0.1
numexpr==2.10.2
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.20.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.12
overrides==7.7.0
packaging==24.0
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pbkdf2==1.3
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
PennyLane==0.38.0
PennyLane_Lightning==0.38.0
peppercorn==0.6
phe==1.5.0
pillow==10.2.0
platformdirs==4.2.0
posthog==3.7.4
pretrainedmodels==0.7.4
proglog==0.1.10
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==3.19.6
prov==2.0.1
psutil==5.9.8
pure-eval==0.2.2
pyaes==1.6.1
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pydot==3.0.2
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.17.2
Pympler==1.1
PyMsgBox==1.0.9
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.0
qrcode==7.4.2
rdflib==6.3.2
redis==3.5.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
rustworkx==0.15.1
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
segno==1.6.1
sentence-transformers==3.3.1
shellescape==3.8.1
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
streamlit==1.22.0
sympy==1.13.1
tenacity==8.3.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.8.30
tiktoken==0.4.0
timm==0.9.7
tinyec==0.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.5.1
torch-geometric==2.6.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
traits==6.3.2
transformers==4.46.3
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.0
validators==0.34.0
vine==5.1.0
virtualenv==20.25.1
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.2
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
